[ADD+FIX] mejorando rutas back y alguna cosa mas del front
