+++ src/api/controllers/user.py
@@ -12,7 +12,7 @@ def create_user(body):
 
         claves_user = body.keys()
 
-        if not "email" in claves_user or not "password" in claves_user or not "name" in claves_user or not "lastName" in claves_user or not "address" in claves_user or not "province" in claves_user or not "postalCode" in claves_user or not "phone" in claves_user or not "country" in claves_user or not "birthdate" in claves_user:
+        if not "email" in claves_user or not "password" in claves_user or not "name" in claves_user or not "lastName" in claves_user or not "address" in claves_user or not "province" in claves_user or not "postalCode" in claves_user or not "phone" in claves_user or not "country" in claves_user or not "birthdate" in claves_user:           # or not "latitude" in claves_user or not "longitude" in claves_user        
             return {"code": 400, "msg": "Missing data in the forms"}
 
 
@@ -29,6 +29,9 @@ def create_user(body):
             country = body["country"], 
             birthdate = body["birthdate"],
             is_active = True)
+
+            # latitude = int(body["latitude"]),
+            # longitude = body["longitude"], 
 
         db.session.add(new_user)
         db.session.commit()
@@ -131,6 +134,9 @@ def update_user(body, id):
         user.birthdate = body["birthdate"]
         user.aboutMe = body["aboutMe"]
         user.is_active = True
+
+        # latitude = int(body["latitude"])
+        # longitude = body["longitude"]
 
         # if body["userPhoto"]:
         #     cloudinary.uploader.upload(user.name + ".mp4",
