+++ src/front/js/constants/RutaBack.jsx
@@ -1,8 +1,8 @@
 
-
-// CAMBIAR CADA VEZ QUE SE LEVANTE EL SERVER DE BACK, YA QUE LA URL CAMBIA
+// YA NO HACE FALTA CAMBIAR, LA COGE SIEMPRE DE .ENV AL CREAR EL SERVER DEL BACK
 
 export const RUTABACK = process.env.BACKEND_URL;
 
 
+// CAMBIAR CADA VEZ QUE SE LEVANTE EL SERVER DE BACK, YA QUE LA URL CAMBIA
 // export const RUTABACK = "https://3001-giousned-proyectfinalco-n2vh2xmouju.ws-eu96b.gitpod.io";
